Move import/export actions to menu
