+++ src/jsMain/kotlin/CodesScreen.kt
@@ -1,14 +1,7 @@
 import dev.fritz2.core.*
-import kotlinx.browser.document
 import kotlinx.browser.window
 import kotlinx.coroutines.flow.map
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
-import org.w3c.dom.HTMLAnchorElement
-import org.w3c.dom.HTMLInputElement
 import org.w3c.dom.HTMLElement
-import org.w3c.files.FileReader
 import qr.QrData
 import qr.QrType
 import qr.SavedQrCode
@@ -22,8 +15,7 @@ import sortable.SortableEvent
 import sortable.sortableOptions
 
 fun RenderContext.codesScreen(
-    savedCodesStore: Store<List<SavedQrCode>>,
-    json: Json
+    savedCodesStore: Store<List<SavedQrCode>>
 ) {
     val formStore = storeOf(QrForm())
     val editingStore = storeOf(false)
@@ -54,53 +46,6 @@ fun RenderContext.codesScreen(
                     clicks handledBy {
                         formStore.update(QrForm())
                         editingStore.update(true)
-                    }
-                }
-
-                val fileInput = input("hidden") {
-                    type("file")
-                    accept(".json")
-                }.apply {
-                    domNode.addEventListener("change", { event ->
-                        val inputElement = event.target as HTMLInputElement
-                        val file = inputElement.files?.item(0)
-                        if (file != null) {
-                            val reader = FileReader()
-                            reader.onload = { loadEvent ->
-                                val text = (loadEvent.target as FileReader).result as String
-                                try {
-                                    val list = json.decodeFromString<List<SavedQrCode>>(text)
-                                        .map { if (it.name.isBlank()) it.copy(name = it.text) else it }
-                                    savedCodesStore.update(list)
-                                } catch (e: Throwable) {
-                                    window.alert(getTranslationString(DefaultLangStrings.InvalidJson))
-                                }
-                                inputElement.value = ""
-                            }
-                            reader.readAsText(file)
-                        }
-                    })
-                }
-
-                button("btn btn-secondary btn-sm w-24 flex items-center gap-1") {
-                    iconArrowUpTray()
-                    translate(DefaultLangStrings.Import)
-                    clicks handledBy {
-                        fileInput.domNode.click()
-                    }
-                }
-                button("btn btn-secondary btn-sm w-24 flex items-center gap-1") {
-                    iconArrowDownTray()
-                    translate(DefaultLangStrings.Export)
-                    clicks handledBy {
-                        val txt = json.encodeToString(savedCodesStore.current)
-                        val encoded = js("encodeURIComponent")(txt) as String
-                        val link = document.createElement("a") as HTMLAnchorElement
-                        link.href = "data:application/json;charset=utf-8,$encoded"
-                        link.download = "codes.json"
-                        document.body?.appendChild(link)
-                        link.click()
-                        document.body?.removeChild(link)
                     }
                 }
             }
